Port MSVC builds from DDA (but without localization for now) (#1392)
* Cataclysm Windows CI definition.

* Ignore data (at least for now) in msvc compile

* Ignore lang subdirectory for MSVC

* Ignore more paths for MSVC

* Put data, tests together at end

These will be removed eventually.

* Build and run the tests in the Windows build

* Compile .mo files on Windows CI

* Use Makefile to compile MO on Windows CI

* Add native msvc builds to release.yml

* Enable GitHub Actions problem matchers in MSVC build (#52138)

* Only run basic build on draft PRs (#53508)

* Simplify windows workflows for latest run-vcpkg release. (#53717)

Co-authored-by: akrieger <akrieger@user.noreply.github.com>

* Set ZSTD_CLEVEL=17 in jobs using GHA caching for space wins. (#54118)

* Eliminate concurrency of expensive builds (#54336)

* Eliminate concurrency of msvc build
* Kill old clang tidy builds
* Kill old object creator builds

* Bump windows builds cachekeys. (#54726)

Co-authored-by: akrieger <akrieger@user.noreply.github.com>

* BN-ify .github directory

* Github workflow for ObjectCreator on Windows

This is a pretty bog standard looking windows workflow. The special
sauce is apparently in the `run-vcpkg` action, which has some magic in
it to cache the resulting built artifacts and automatically restore them
in subsequent runs. This is good, because the vcpkg build is very slow.
There's some subtleties in it though. We have to manually add the
msbuild integration's output paths to the list of cached folders. The
cache also includes pretty much _all_ of the vcpkg folder except build
outputs, which is problematic because it means we can't modify the
builtin triplets on the fly.

Why do we want to modify the builtin triplets? What are the triplets,
even? Triplets are basically short descriptive strings which tell vcpkg
what 'kind' of build you want - eg. `x86-windows`, `x64-windows`,
`x64-windows-static`, whatever. We want `x64-windows-static`, because
dll deployment is a cluster. Regardless, what these triplets map to are
essentially tiny CMake files which set some common configuration for all
the deps that vcpkg builds. These CMake files are usually checked into
vcpkg's repo. If we wanted to set some configuration 'globally' for all
deps, we'd have to insert it into the triplet files directly if there
wasn't some command line argument or environment variable we could use.

So why do we care? By default, vcpkg will build both debug and release
versions of all deps, and this takes a large amount of disk space. Too
much disk space, in fact, for the workers to still complete successfully
(and upload to cache if needed). One way to work around this would be to
insert `set(VCPKG_BUILD_TYPE release)` into every triplet file. This
works, but it breaks `run-vcpkg`'s caching. More specifically, because
`run-vcpkg` caches the entire vcpkg checkout including the triplets, on
subsequent runs if we were to naively insert `set(VCPKG_BUILD_TYPE release)`
again it would simply append after the entry we already had. vcpkg
itself though has extra checks to detect if packages need to be rebuilt,
including hashing the triplet files. If we just keep appending the
setting we will keep changing the triplets and busting vcpkg's cache.
Instead we use 'triplet overlays' which are essentially 'triplet include
paths' that tell vcpkg where to look for triplet files. We'll point
vcpkg at a folder in our repo where we've checked in shadow copies of
the triplets we care about which include the setting automatically, and
will thus be consistent on every run, and no one's cache will be busted.

* Fix vcpkg configuration for release.yml

* Try without gettext

* Try without translations for now

* Rename JsonFormatter

* Add vcpkg-static to JsonFormatter again

* Remove ccpp.yml

* Less translation...

* Matchers not implemented yet

Co-authored-by: Andrew Krieger <akrieger@users.noreply.github.com>
Co-authored-by: actual-nh <74678550+actual-nh@users.noreply.github.com>
Co-authored-by: John Bytheway <52664+jbytheway@users.noreply.github.com>
Co-authored-by: Binrui Dong <brett.browning.dong@gmail.com>
Co-authored-by: Kevin Granade <kevin.granade@gmail.com>
Co-authored-by: akrieger <akrieger@user.noreply.github.com>