ci: add problem matchers (#2660)
* feat: display multi-line error messages with github action

Co-authored-by: Qrox <qrox@sina.com>

* Add debugmsg problem matcher

This problem matcher is intended to highlight errors coming from hitting
debugmsg in CI.

Co-authored-by: John Bytheway <jbytheway@gmail.com>

* Make path appear in __FILE__ for tests

The DebugLog printing of filenames relies on __FILE__.  This is based on
the filename passed to the compiler.  For tests, that lacked a directory
name.  So, __FILE__ lacked the context for the filename, which meant the
GitHub problem matcher couldn't find the file causing the error.

Add '../tests/' to the filename passed to the compiler to work around
this.

Co-authored-by: John Bytheway <jbytheway@gmail.com>

* Add Catch2 problem matcher

This is intended to highlight test failures in the Catch2 tests as
GitHub annotations.

Co-authored-by: John Bytheway <jbytheway@gmail.com>

* ci: add msvc problem matchers

uses https://github.com/marketplace/actions/msvc-problem-matcher

* style: false positive for `\r`

---------

Co-authored-by: Qrox <qrox@sina.com>
Co-authored-by: John Bytheway <jbytheway@gmail.com>